Only show boot output when --show-boot passed
We want to see user space output all the time, but only show boot output
when the `--show-boot` argument is provided.

Put the kernel messages on tty0 and have other messages go to tty1 when we
don't want to see the boot messages, if we do send everything to tty0.

Signed-off-by: Martyn Welch <martyn.welch@collabora.com>